refactor(modularity): eliminate utils cross-module re-export
Remove cross-module re-exports from vibe3.utils to vibe3.clients.runtime_assets:
- resolve_prompt_config
- resolve_runtime_asset
- runtime_assets_root

Changes:
- utils/__init__.py: Remove cross-module import block and __all__ entries
- roles/manager.py: Import runtime_assets_root from vibe3.clients
- roles/run_command.py: Import resolve_runtime_asset from vibe3.clients

This eliminates the indirection layer where utils re-exported symbols from
clients, improving module boundary clarity.

Ref: #2279
